LayoutTests/imported/w3c/web-platform-tests/media-source/mediasource-correct-frames-expected.txt

CONSOLE MESSAGE: Error: assert_true: The note being played matches the expected one (boxes lit: 0, 440.0 Hz), found ~-5.4 Hz expected true got false
CONSOLE MESSAGE: Error: assert_equals: Num of lit boxes: expected 1 but got 0
CONSOLE MESSAGE: Error: assert_equals: Num of lit boxes: expected 1 but got 0
CONSOLE MESSAGE: Error: assert_equals: Num of lit boxes: expected 1 but got 0
CONSOLE MESSAGE: Error: assert_equals: Num of lit boxes: expected 1 but got 0
CONSOLE MESSAGE: Error: assert_equals: Num of lit boxes: expected 1 but got 0
CONSOLE MESSAGE: Error: assert_equals: Num of lit boxes: expected 1 but got 0
CONSOLE MESSAGE: Error: assert_equals: Num of lit boxes: expected 1 but got 0
CONSOLE MESSAGE: Error: assert_equals: Num of lit boxes: expected 1 but got 0
CONSOLE MESSAGE: Error: assert_equals: Num of lit boxes: expected 1 but got 0
CONSOLE MESSAGE: Error: assert_equals: Num of lit boxes: expected 1 but got 0
CONSOLE MESSAGE: Error: assert_equals: Num of lit boxes: expected 1 but got 0
CONSOLE MESSAGE: Error: assert_equals: Num of lit boxes: expected 1 but got 0
CONSOLE MESSAGE: Error: assert_equals: Num of lit boxes: expected 1 but got 0
CONSOLE MESSAGE: Error: assert_equals: Num of lit boxes: expected 1 but got 0
CONSOLE MESSAGE: Error: assert_equals: Num of lit boxes: expected 1 but got 0
CONSOLE MESSAGE: Error: assert_equals: Num of lit boxes: expected 1 but got 0
CONSOLE MESSAGE: Error: assert_equals: Num of lit boxes: expected 1 but got 0
CONSOLE MESSAGE: Error: assert_equals: Num of lit boxes: expected 1 but got 0
CONSOLE MESSAGE: Error: assert_equals: Num of lit boxes: expected 1 but got 0
CONSOLE MESSAGE: Error: assert_equals: Num of lit boxes: expected 1 but got 0
CONSOLE MESSAGE: Error: assert_equals: Num of lit boxes: expected 1 but got 0
CONSOLE MESSAGE: Error: assert_equals: Num of lit boxes: expected 1 but got 0
CONSOLE MESSAGE: Error: assert_equals: Num of lit boxes: expected 1 but got 0
CONSOLE MESSAGE: Error: assert_equals: Num of lit boxes: expected 1 but got 0
CONSOLE MESSAGE: Error: assert_equals: Num of lit boxes: expected 1 but got 0
CONSOLE MESSAGE: Error: assert_equals: Num of lit boxes: expected 1 but got 0
CONSOLE MESSAGE: Error: assert_equals: Num of lit boxes: expected 1 but got 0
CONSOLE MESSAGE: Error: assert_equals: Boxes lit at video ended event expected 7 but got 0

Harness Error (FAIL), message = Error: assert_equals: Boxes lit at video ended event expected 7 but got 0

TIMEOUT Test the expected frames are played at the expected times Test timed out

Harness Error (FAIL), message = Error: assert_equals: Boxes lit at video ended event expected 7 but got 0

TIMEOUT Test the expected frames are played at the expected times Test timed out
